+++ plugins/postcss-nesting/src/lib/atrule-within-atrule.ts
@@ -1,16 +1,12 @@
+import type { AtRule } from 'postcss';
 import cleanupParent from './cleanup-parent.js';
 import mergeParams from './merge-params.js';
 import shiftNodesBeforeParent from './shift-nodes-before-parent.js';
 import validAtrules from './valid-atrules.js';
 
-/*
- * DEPRECATED: In v7.0.0 these features will be removed as they are not part of
- * the nesting proposal.
- */
-
-export default function transformAtruleWithinAtrule(node) {
+export default function transformAtruleWithinAtrule(node: AtRule, parent: AtRule) {
 	// move previous siblings and the node to before the parent
-	const parent = shiftNodesBeforeParent(node);
+	shiftNodesBeforeParent(node, parent);
 
 	// update the params of the node to be merged with the parent
 	node.params = mergeParams(parent.params, node.params);
@@ -19,4 +15,7 @@ export default function transformAtruleWithinAtrule(node) {
 	cleanupParent(parent);
 }
 
-export const isAtruleWithinAtrule = (node) => node.type === 'atrule' && validAtrules.includes(node.name) && Object(node.parent).type === 'atrule' && node.name === node.parent.name;
+export function isAtruleWithinAtrule(node: AtRule, parent: AtRule) {
+	return validAtrules.includes(node.name) &&
+		node.name === parent.name;
+}