+++ plugins/postcss-cascade-layers/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes to PostCSS Cascade Layers
 
-### Unreleased
+### 1.1.0 (September 14, 2022)
 
 - Add support for `@scope` and `@container` as parent rules for `@layer`